refactor: rename module pdf_content.py to document_content.py
