+++ src/fennflow/files/media/__init__.py
@@ -13,9 +13,9 @@ __all__ = [
 from .audio_content import AudioContent
 from .base import BaseContent
 from .binary_content import BinaryContent
+from .document_content import DocumentContent
 from .image_content import ImageContent
 from .json_content import JsonContent
-from .pdf_content import DocumentContent
 from .text_content import TextContent
 from .url_content import UrlContent
 from .video_content import VideoContent
